# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UniPath MIS - Genomics & Bioinformatics Project Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: serviceHead
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    put:
      operationId: updateService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────────────────────────────────────────
  # SCIENTISTS
  # ──────────────────────────────────────────────────────────────────────────
...
  /territories:
    get:
      operationId: listTerritories
      tags: [territories]
      summary: List all territories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Territory"
    post:
      operationId: createTerritory
      tags: [territories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerritoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Territory"
  /territories/{id}:
    put:
      operationId: updateTerritory
      tags: [territories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerritoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Territory"
    delete:
      operationId: deleteTerritory
      tags: [territories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────────────────────────────────────────
  # SALES PERSONS
  # ──────────────────────────────────────────────────────────────────────────
  /sales-persons:
    get:
      operationId: listSalesPersons
      tags: [sales-persons]
      summary: List all sales persons
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesPerson"
    post:
      operationId: createSalesPerson
      tags: [sales-persons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesPersonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesPerson"
  /sales-persons/{id}:
    put:
      operationId: updateSalesPerson
      tags: [sales-persons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesPersonInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesPerson"
    delete:
      operationId: deleteSalesPerson
      tags: [sales-persons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────────────────────────────────────────────────────────────
  # PROJECTS
  # ──────────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Service:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        serviceHead:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
    ServiceInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        serviceHead:
          type: string
        category:
          type: string
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        serviceHead:
          type: string
        category:
          type: string
...
    Territory:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    TerritoryInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    SalesPerson:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        territoryId:
          type: integer
        territoryName:
          type: string
        createdAt:
          type: string
          format: date-time
    SalesPersonInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
        territoryId:
          type: integer
...