# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * UniPath MIS - Genomics & Bioinformatics Project Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  serviceHead?: string;
  category?: string;
  createdAt?: Date;
}